Add interface to add state machines
This adds the ability to create and update state machines for resources in MORYX
projects.

Right now, it is possible to use

    $ moryx add states <ResourceName> --states "<State1>, <...>"`

to create a `StateBase` class for the provided resource together with all states
passed to the `--states` option. It also updates the resource to implement
`IStateContext`

If the command gets executed on a resource that already has a state machine, the
related classes get updated accordingly.