Create a basic template config for creating a new project
The idea is to have a template description inside source repositories
(`.moryxtpl`) that define, which files are going to be replaced on a
specific command and what strings will be replaced (in filenames and file
content).

So that it would be easy to customize creation of applications without
needing to how the CLI processes things internally.