Introduce 'remotes' to manage 'profiles'
From the beginning, `Profiles` where introduced to make it possible to choose
between several source repositories. Back then, the main use case of profiles
was to store the information from 'which' template a project was generated
initially and to be able to update those templates when needed.

`remotes` now extends this concept by making it possible to 'manage' profiles
by enabling to add and remove profiles during development and changing the
default profile to be used.

The feature is called `remotes` as it is very close to Git's remotes. Under
the hood it will stick with `Profiles` until the feature gets to a more mature
state.